Update Attendance Being Processed

+++ db/migrate/20180223060805_create_attendances.rb
@@ -1,9 +1,9 @@
 class CreateAttendances < ActiveRecord::Migration[5.1]
   def change
     create_table :attendances do |t|
-      t.integer :present
-      t.integer :absent
-      t.integer :total
+      t.float :present
+      t.float :absent
+      t.float :total
       t.references :student, foreign_key: true
  
       t.timestamps
